SmartHaven/SmartHavenWeb/requirements.txt
Adafruit-Blinka==8.68.0
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-ina219==3.4.31
adafruit-circuitpython-neopixel==6.3.21
adafruit-circuitpython-pixelbuf==2.0.10
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.85.0
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
attrs==25.4.0
binho-host-adapter==0.1.6
blinker==1.9.0
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
Flask==3.1.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
influxdb==5.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mfrc522-python==0.0.7
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.6.1
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
propcache==0.4.1
psutil==7.2.1
pubnub==10.5.0
pycparser==2.23
pycryptodomex==3.23.0
pyftdi==0.57.1
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.3
requests==2.32.5
RPi.bme280==0.2.4
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
smbus2==0.5.0
sounddevice==0.5.3
spidev==3.8
sympy==1.14.0
sysv_ipc==1.1.0
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
ultralytics==8.4.6
ultralytics-thop==2.0.18
urllib3==2.5.0
Werkzeug==3.1.4
yarl==1.22.0
